refactor(#130): types 디렉토리 재사용
