+++ apps/web/src/app/auth/steps/EmailStep.tsx
@@ -1,10 +1,10 @@
 import ContinueWithGoogle from '@/components/signin/ContinueWithGoogle/ContinueWithGoogle';
+import type { EmailStepProps } from '@/types';
 import styled from '@emotion/styled';
 import { tokens } from '@horizon/tokens';
 import { Text } from '@horizon/ui';
 import { Flexbox } from '@horizon/utils';
 import SecondaryAction from '../../../components/signup/SecondaryAction/SecondaryAction';
-import type { EmailStepProps } from '../../../types/types';
 
 const EmailStep = ({
   email,
