+++ apps/web/src/app/auth/steps/PasswordStep.tsx
@@ -1,8 +1,8 @@
+import type { PasswordStepProps } from '@/types';
 import styled from '@emotion/styled';
 import { tokens } from '@horizon/tokens';
 import { Text, TextField } from '@horizon/ui';
 import { Flexbox } from '@horizon/utils';
-import type { PasswordStepProps } from '../../../types/types';
 
 const PasswordStep = ({
   password,
