+++ apps/web/src/hooks/auth/signup/useSignUpHandlers.ts
@@ -1,7 +1,7 @@
+import type { SignUpData, SignUpStep } from '@/types';
 import { useRouter } from 'next/navigation';
 import type React from 'react';
 import { useState } from 'react';
-import type { SignUpData, SignUpStep } from '../../../types/types';
 import { useRegister } from './useSignUp';
 import { useSignUpValidation } from './useSignUpValidation';
