+++ apps/web/src/types/index.ts
@@ -1,0 +1,10 @@
+// Auth types
+export type {
+  SignUpStep,
+  SignUpData,
+  StepProps,
+  EmailStepProps,
+  VerificationStepProps,
+  PasswordStepProps,
+  UsernameStepProps,
+} from './auth/signup/types';
